+++ eras/dijkstra/impl/CHANGELOG.md
@@ -97,6 +97,7 @@
 
 ### `testlib`
 
+* Add `txWithSubTxs` and `submitFailingSubTx`
 * Preserve explicitly supplied redeemers when fixing up subtransactions.
 * Add `InjectRuleFailure "LEDGER" DijkstraSubUtxowPredFailure era` as a superclass of `DijkstraEraImp`
 * Add `mkTopTxWithSubTxs`, `traverseSubTxs` and `withPostFixupSubTxs`